requirements.txt
streamlit>=1.56.0
openpyxl>=3.1.5
Pillow>=12.2.0
